Clean up on North Macedonian dates
