+++ docs/northmacedonia.md
@@ -1,13 +1,5 @@
 ### North Macedonian dates
 
-- `isOrthodoxChristmasEve
-- `isOrthodoxChristmasDay
-- `isDayOffForOrthodoxChristmasDay
-- `isOrthodoxEpiphany
-- `isNorthMacedonianLaborDay
-- `isNorthMacedonianRepublicDay
-- `isFeastOfTheAssumptionOfMaryOrthodox
-- `isNorthMacedonianIndependenceDay
-- `isNorthMacedonianDayOfPeoplesUprising
-- `isDayOfTheMacedonianRevolutionaryStruggle
-- `isNorthMacedonianSaintKlimentOhridskisDay
+- `isNorthMacedonianRepublicDay`
+- `isNorthMacedonianIndependenceDay`
+- `isNorthMacedonianDayOfPeoplesUprising`
