+++ tests/CanadianHolidaysTest.php
@@ -5,86 +5,6 @@ use Illuminate\Support\Carbon;
 class CanadianHolidaysTest extends TestCase
 {
     /**
-     * @dataProvider provideNewYearsDayData
-     */
-    public function test_it_knows_new_years_day($date, $validity)
-    {
-        $date = Carbon::parse($date);
-
-        $this->assertSame($validity, $date->isNewYearsDay());
-    }
-
-    public function provideNewYearsDayData()
-    {
-        return [
-            '1963-11-22' => ['1963-11-22', false],
-            '1970-01-01' => ['1970-01-01', true],
-            '1982-01-01' => ['1982-01-01', true],
-            '1999-12-31' => ['1999-12-31', false],
-            '2020-01-01' => ['2020-01-01', true],
-            '2020-01-02' => ['2020-01-02', false],
-            '2050-01-01' => ['2050-01-01', true],
-            '2038-01-19' => ['2038-01-19', false],
-        ];
-    }
-
-    /**
-     * @dataProvider provideEasterSundayData
-     */
-    public function test_it_knows_easter_sunday($date, $validity)
-    {
-        $date = Carbon::parse($date);
-
-        $this->assertSame($validity, $date->isEasterSunday());
-    }
-
-    public function provideEasterSundayData()
-    {
-        return [
-            '1905-04-23' => ['1905-04-23', true],
-            '1970-03-28' => ['1970-03-28', false],
-            '1970-03-28' => ['1970-03-28', false],
-            '1970-03-29' => ['1970-03-29', true],
-            '1970-03-30' => ['1970-03-30', false],
-            '1990-04-14' => ['1990-04-14', false],
-            '1990-04-15' => ['1990-04-15', true],
-            '1990-04-16' => ['1990-04-16', false],
-            '2020-04-12' => ['2020-04-12', true],
-            '2020-04-13' => ['2020-04-13', false],
-            '2050-04-10' => ['2050-04-10', true],
-            '2050-04-11' => ['2050-04-11', false],
-        ];
-    }
-
-    /**
-     * @dataProvider provideGoodFridayData
-     */
-    public function test_it_knows_good_friday($date, $validity)
-    {
-        $date = Carbon::parse($date);
-
-        $this->assertSame($validity, $date->isGoodFriday());
-    }
-
-    public function provideGoodFridayData()
-    {
-        return [
-            '1905-04-21' => ['1905-04-21', true],
-            '1970-03-26' => ['1970-03-26', false],
-            '1970-03-27' => ['1970-03-27', true],
-            '1970-03-28' => ['1970-03-28', false],
-            '1970-03-30' => ['1970-03-30', false],
-            '1990-04-14' => ['1990-04-14', false],
-            '1990-04-13' => ['1990-04-13', true],
-            '1990-04-16' => ['1990-04-16', false],
-            '2020-04-10' => ['2020-04-10', true],
-            '2020-04-13' => ['2020-04-13', false],
-            '2050-04-08' => ['2050-04-08', true],
-            '2050-04-11' => ['2050-04-11', false],
-        ];
-    }
-
-    /**
      * @dataProvider provideFamilyDayData
      */
     public function test_it_knows_family_day($date, $validity)
@@ -230,33 +150,6 @@ class CanadianHolidaysTest extends TestCase
             '2020-11-11' => ['2020-11-11', true],
             '2020-11-12' => ['2020-11-12', false],
             '2050-11-11' => ['2050-11-11', true],
-        ];
-    }
-
-    /**
-     * @dataProvider provideChristmasDayData
-     */
-    public function test_it_knows_christmas_day($date, $validity)
-    {
-        $date = Carbon::parse($date);
-
-        $this->assertSame($validity, $date->isChristmasDay());
-    }
-
-    public function provideChristmasDayData()
-    {
-        return [
-            '1900-12-25' => ['1900-12-25', true],
-            '1900-12-26' => ['1900-12-26', false],
-            '1920-12-25' => ['1920-12-25', true],
-            '1950-12-25' => ['1950-12-25', true],
-            '1980-12-24' => ['1980-12-24', false],
-            '1980-12-25' => ['1980-12-25', true],
-            '1980-12-26' => ['1980-12-26', false],
-            '2020-12-25' => ['2020-12-25', true],
-            '2021-12-25' => ['2021-12-25', true],
-            '2021-12-26' => ['2021-12-26', false],
-            '2050-12-25' => ['2050-12-25', true],
         ];
     }
